Extract oscilloscope into separate module
