+++ src/components/WorkspacePanel.vue
@@ -19,7 +19,7 @@
                 <div class=""><VCFModule /></div>
                 <div class=""><SynthPanel /></div>
             </div>
-            <div class="grid grid-cols-2 sm:grid-cols-5 gap-4 mx-4">
+            <div class="grid grid-cols-2 sm:grid-cols-6 gap-4 mx-4">
                 <div class=""><MixerModule /></div>
                 <div class=""><EnvelopeGenerator /></div>
                 <div class="flex flex-col gap-4">
@@ -27,6 +27,7 @@
                 </div>
                 <div class=""><VCAModule /></div>
                 <div class=""><MasterOutput /></div>
+                <div class=""><Oscilloscope /></div>
             </div>
         </div>
 
@@ -50,6 +51,7 @@ import MixerModule from './modules/MixerModule.vue';
 import SynthPanel from './modules/SynthPanel.vue';
 import PatchCables from './PatchCables.vue';
 import MasterOutput from "./modules/MasterOutput.vue";
+import Oscilloscope from "./modules/Oscilloscope.vue";
 
 const audioReady = ref(false);
 const synth = useSynthStore();
